requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
markdownify>=0.11.0
schedule>=1.2.0
python-dotenv>=1.0.0